+++ libcst/codemod/_cli.py
@@ -16,8 +16,8 @@ import time
 import traceback
 from dataclasses import dataclass, replace
 from multiprocessing import cpu_count, Pool
-from pathlib import Path, PurePath
-from typing import Any, AnyStr, cast, Dict, List, Optional, Sequence, Tuple, Union
+from pathlib import Path
+from typing import Any, AnyStr, cast, Dict, List, Optional, Sequence, Union
 
 from libcst import parse_module, PartialParserConfig
 from libcst.codemod._codemod import Codemod
@@ -32,6 +32,7 @@ from libcst.codemod._runner import (
     TransformSkip,
     TransformSuccess,
 )
+from libcst.helpers import calculate_module_and_package
 from libcst.metadata import FullRepoManager
 
 _DEFAULT_GENERATED_CODE_MARKER: str = f"@gen{''}erated"
@@ -184,35 +185,6 @@ def exec_transform_with_prettyprint(
     return maybe_code
 
 
-def _calculate_module_and_package(
-    repo_root: Optional[str], filename: str
-) -> Optional[Tuple[str, str]]:
-    # Given an absolute repo_root and an absolute filename, calculate the
-    # python module name for the file.
-    if repo_root is None:
-        # We don't have a repo root, so this is impossible to calculate.
-        return None
-
-    try:
-        relative_filename = PurePath(filename).relative_to(repo_root)
-    except ValueError:
-        # This file seems to be out of the repo root.
-        return None
-
-    # get rid of extension
-    relative_filename = relative_filename.with_suffix("")
-
-    # handle special cases
-    if relative_filename.stem in ["__init__", "__main__"]:
-        relative_filename = relative_filename.parent
-        package = module = ".".join(relative_filename.parts)
-    else:
-        module = ".".join(relative_filename.parts)
-        package = ".".join(relative_filename.parts[:-1])
-
-    return module, package
-
-
 @dataclass(frozen=True)
 class ExecutionResult:
     # File we have results for
@@ -269,11 +241,6 @@ def _execute_transform(  # noqa: C901
                 ),
             )
 
-        # attempt to work out the module and package name for this file
-        full_module_name, full_package_name = _calculate_module_and_package(
-            config.repo_root, filename
-        ) or (None, None)
-
         # Somewhat gross hack to provide the filename in the transform's context.
         # We do this after the fork so that a context that was initialized with
         # some defaults before calling parallel_exec_transform_with_prettyprint
@@ -281,10 +248,19 @@ def _execute_transform(  # noqa: C901
         transformer.context = replace(
             transformer.context,
             filename=filename,
-            full_module_name=full_module_name,
-            full_package_name=full_package_name,
             scratch={},
         )
+
+        # attempt to work out the module and package name for this file
+        module_name_and_package = calculate_module_and_package(
+            config.repo_root, filename
+        )
+        if module_name_and_package is not None:
+            transformer.context = replace(
+                transformer.context,
+                full_module_name=module_name_and_package.name,
+                full_package_name=module_name_and_package.package,
+            )
 
         # Run the transform, bail if we failed or if we aren't formatting code
         try:
